Crate mithril_cardano_node_internal_database

This crate provides components to read the files of a Cardano node internal database, such as immutable files and ledger state snapshots, and compute digests from them. It includes observers, digesters and various helpers for file system operations.

Modules

digesters

Tools to compute Mithril digests from a Cardano node database.

entities

Entities related to a Cardano node internal database.

signable_builder

Signable Builders for Cardano node internal database artifacts

test

Test utilities.

Structs

ImmutableFileSystemObserver

An ImmutableFileObserver using the filesystem.

Enums

ImmutableFileObserverError

ImmutableFileObserver related errors.

Constants

IMMUTABLE_DIR

Directory name for the immutable files.

LEDGER_DIR

Directory name for the ledger files.

VOLATILE_DIR

Directory name for the volatile files.

Traits

ImmutableFileObserver

Retrieve data on ImmutableFile from a cardano database.

Functions

immutable_trio_names

Returns the names of the files that compose an immutable trio.