mithril_cardano_node_internal_database

Module digesters

Tools to compute Mithril digests from a Cardano node database.

Modules

cache

Cache provider to accelerate super::CardanoImmutableDigester computation by storing and reusing crate::entities::ImmutableFile digests.

Structs

CardanoImmutableDigester

A digester working directly on a Cardano DB immutables files

ComputedImmutablesDigests

Computed immutables digests

Enums

ImmutableDigesterError

ImmutableDigester related Errors.

Traits

ImmutableDigester

A digester that can compute the digest used for mithril signatures